working xsp v1 TODO: support code for EID usage TODO: more testing
git-svn-id: https://damsl.cs.indiana.edu/svn/xsp/trunk@144 6dc24a1a-a86e-4cf8-a923-33330a3f456b

+++ src/libxsp_client/libxsp_client.h
@@ -34,8 +34,8 @@ int xsp_set_session_socket(libxspSess *sess, int new_sd);
 int xsp_set_session_connected(libxspSess *sess);
 int xsp_send_ping(libxspSess *sess);
 int xsp_recv_ping(libxspSess *sess);
-int xsp_send_msg(libxspSess *sess, const void *buf, size_t len, int opt_type);
-int xsp_recv_msg(libxspSess *sess, void **ret_buf, int *len, int *ret_type);
+int xsp_send_msg(libxspSess *sess, const void *buf, uint64_t len, int opt_type);
+int xsp_recv_msg(libxspSess *sess, void **ret_buf, uint64_t *len, int *ret_type);
 int xsp_sesscmp(libxspSess *s1, libxspSess *s2);
 
 #endif
